# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sayq POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
  /semi-finished:
    get:
      operationId: listSemiFinished
      tags: [semiFinished]
      summary: List semi-finished goods with live stock
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of semi-finished goods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SemiFinished"
    post:
      operationId: createSemiFinished
      tags: [semiFinished]
      summary: Create a semi-finished good
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemiFinishedInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemiFinished"
  /semi-finished/{id}:
    patch:
      operationId: updateSemiFinished
      tags: [semiFinished]
      summary: Update a semi-finished good
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemiFinishedUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemiFinished"
    delete:
      operationId: deleteSemiFinished
      tags: [semiFinished]
      summary: Delete a semi-finished good
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /semi-finished/{id}/produce:
    post:
      operationId: produceSemiFinished
      tags: [semiFinished]
      summary: Record production of a semi-finished good (deducts ingredients, computes COGS)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduceInput"
      responses:
        "200":
          description: Produced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemiFinished"
  /recipes:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get the BOM components for a parent
      parameters:
        - name: parentType
          in: query
          required: true
          schema:
            type: string
            enum: [product, semi_finished]
        - name: parentId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeComponent"
    put:
      operationId: setRecipe
      tags: [recipes]
      summary: Replace the BOM for a parent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRecipeInput"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeComponent"
...
components:
  schemas:
...
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderInput:
      type: object
      required: [items, paymentMethod, amountPaid]
      properties:
        branchId:
          type: ["integer", "null"]
        cashierName:
          type: string
        cashierId:
          type: ["integer", "null"]
        paymentMethod:
          type: string
          enum: [cash, card, qris]
        amountPaid:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, priceAtSale, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        priceAtSale:
          type: number
        subtotal:
          type: number
    Order:
      type: object
      required: [id, total, paymentMethod, status, createdAt]
      properties:
        id:
          type: integer
        cashierName:
          type: ["string", "null"]
        cashierId:
          type: ["integer", "null"]
        total:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        itemCount:
          type: integer
    OrderDetail:
      type: object
      required: [id, total, paymentMethod, status, createdAt, items]
      properties:
        id:
          type: integer
        cashierName:
          type: ["string", "null"]
        cashierId:
          type: ["integer", "null"]
        total:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...
    SemiFinished:
      type: object
      required: [id, branchId, name, unit, costPricePerUnit, currentStock]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        name:
          type: string
        unit:
          type: string
        costPricePerUnit:
          type: number
        currentStock:
          type: number
    SemiFinishedInput:
      type: object
      required: [branchId, name, unit]
      properties:
        branchId:
          type: integer
        name:
          type: string
          minLength: 1
        unit:
          type: string
    SemiFinishedUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        unit:
          type: string
    ProduceInput:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: number
          minimum: 0
    RecipeComponent:
      type: object
      required: [componentType, componentId, quantity]
      properties:
        id:
          type: integer
        componentType:
          type: string
          enum: [ingredient, semi_finished]
        componentId:
          type: integer
        componentName:
          type: string
        unit:
          type: string
        quantity:
          type: number
    SetRecipeInput:
      type: object
      required: [parentType, parentId, components]
      properties:
        parentType:
          type: string
          enum: [product, semi_finished]
        parentId:
          type: integer
        components:
          type: array
          items:
            type: object
            required: [componentType, componentId, quantity]
            properties:
              componentType:
                type: string
                enum: [ingredient, semi_finished]
              componentId:
                type: integer
              quantity:
                type: number
...