# lib/api-zod/src/generated/types/semiFinishedInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sayq POS API
 * OpenAPI spec version: 0.1.0
 */

export interface SemiFinishedInput {
  branchId: number;
  /** @minLength 1 */
  name: string;
  unit: string;
}